[FLINK-10242][tests] Split StreamSourceOperatorTest
