+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/operators/StreamSourceOperatorLatencyMetricsTest.java
@@ -20,7 +20,6 @@ package org.apache.flink.streaming.runtime.operators;
 
 import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.accumulators.Accumulator;
-import org.apache.flink.api.common.functions.StoppableFunction;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.runtime.execution.Environment;
 import org.apache.flink.runtime.jobgraph.OperatorID;
@@ -30,10 +29,7 @@ import org.apache.flink.streaming.api.TimeCharacteristic;
 import org.apache.flink.streaming.api.functions.source.SourceFunction;
 import org.apache.flink.streaming.api.graph.StreamConfig;
 import org.apache.flink.streaming.api.operators.Output;
-import org.apache.flink.streaming.api.operators.StoppableStreamSource;
 import org.apache.flink.streaming.api.operators.StreamSource;
-import org.apache.flink.streaming.api.operators.StreamSourceContexts;
-import org.apache.flink.streaming.api.watermark.Watermark;
 import org.apache.flink.streaming.runtime.streamrecord.StreamElement;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 import org.apache.flink.streaming.runtime.streamstatus.StreamStatus;
@@ -54,7 +50,6 @@ import java.util.Collections;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -63,120 +58,7 @@ import static org.mockito.Mockito.when;
  * Tests for {@link StreamSource} operators.
  */
 @SuppressWarnings("serial")
-public class StreamSourceOperatorTest {
-
-	@Test
-	public void testEmitMaxWatermarkForFiniteSource() throws Exception {
-
-		// regular stream source operator
-		StreamSource<String, FiniteSource<String>> operator =
-				new StreamSource<>(new FiniteSource<String>());
-
-		final List<StreamElement> output = new ArrayList<>();
-
-		setupSourceOperator(operator, TimeCharacteristic.EventTime, 0, 0);
-		operator.run(new Object(), mock(StreamStatusMaintainer.class), new CollectorOutput<String>(output));
-
-		assertEquals(1, output.size());
-		assertEquals(Watermark.MAX_WATERMARK, output.get(0));
-	}
-
-	@Test
-	public void testNoMaxWatermarkOnImmediateCancel() throws Exception {
-
-		final List<StreamElement> output = new ArrayList<>();
-
-		// regular stream source operator
-		final StreamSource<String, InfiniteSource<String>> operator =
-				new StreamSource<>(new InfiniteSource<String>());
-
-		setupSourceOperator(operator, TimeCharacteristic.EventTime, 0, 0);
-		operator.cancel();
-
-		// run and exit
-		operator.run(new Object(), mock(StreamStatusMaintainer.class), new CollectorOutput<String>(output));
-
-		assertTrue(output.isEmpty());
-	}
-
-	@Test
-	public void testNoMaxWatermarkOnAsyncCancel() throws Exception {
-
-		final List<StreamElement> output = new ArrayList<>();
-		final Thread runner = Thread.currentThread();
-
-		// regular stream source operator
-		final StreamSource<String, InfiniteSource<String>> operator =
-				new StreamSource<>(new InfiniteSource<String>());
-
-		setupSourceOperator(operator, TimeCharacteristic.EventTime, 0, 0);
-
-		// trigger an async cancel in a bit
-		new Thread("canceler") {
-			@Override
-			public void run() {
-				try {
-					Thread.sleep(200);
-				} catch (InterruptedException ignored) {}
-				operator.cancel();
-				runner.interrupt();
-			}
-		}.start();
-
-		// run and wait to be canceled
-		try {
-			operator.run(new Object(), mock(StreamStatusMaintainer.class), new CollectorOutput<String>(output));
-		}
-		catch (InterruptedException ignored) {}
-
-		assertTrue(output.isEmpty());
-	}
-
-	@Test
-	public void testNoMaxWatermarkOnImmediateStop() throws Exception {
-
-		final List<StreamElement> output = new ArrayList<>();
-
-		// regular stream source operator
-		final StoppableStreamSource<String, InfiniteSource<String>> operator =
-				new StoppableStreamSource<>(new InfiniteSource<String>());
-
-		setupSourceOperator(operator, TimeCharacteristic.EventTime, 0, 0);
-		operator.stop();
-
-		// run and stop
-		operator.run(new Object(), mock(StreamStatusMaintainer.class), new CollectorOutput<String>(output));
-
-		assertTrue(output.isEmpty());
-	}
-
-	@Test
-	public void testNoMaxWatermarkOnAsyncStop() throws Exception {
-
-		final List<StreamElement> output = new ArrayList<>();
-
-		// regular stream source operator
-		final StoppableStreamSource<String, InfiniteSource<String>> operator =
-				new StoppableStreamSource<>(new InfiniteSource<String>());
-
-		setupSourceOperator(operator, TimeCharacteristic.EventTime, 0, 0);
-
-		// trigger an async cancel in a bit
-		new Thread("canceler") {
-			@Override
-			public void run() {
-				try {
-					Thread.sleep(200);
-				} catch (InterruptedException ignored) {}
-				operator.stop();
-			}
-		}.start();
-
-		// run and wait to be stopped
-		operator.run(new Object(), mock(StreamStatusMaintainer.class), new CollectorOutput<String>(output));
-
-		assertTrue(output.isEmpty());
-	}
+public class StreamSourceOperatorLatencyMetricsTest {
 
 	/**
 	 * Test that latency marks are emitted.
@@ -194,10 +76,10 @@ public class StreamSourceOperatorTest {
 
 		// regular stream source operator
 		final StreamSource<Long, ProcessingTimeServiceSource> operator =
-				new StreamSource<>(new ProcessingTimeServiceSource(testProcessingTimeService, processingTimes));
+			new StreamSource<>(new ProcessingTimeServiceSource(testProcessingTimeService, processingTimes));
 
 		// emit latency marks every 10 milliseconds.
-		setupSourceOperator(operator, TimeCharacteristic.EventTime, 0, latencyMarkInterval, testProcessingTimeService);
+		setupSourceOperator(operator, TimeCharacteristic.EventTime, latencyMarkInterval, testProcessingTimeService);
 
 		// run and wait to be stopped
 		operator.run(new Object(), mock(StreamStatusMaintainer.class), new CollectorOutput<Long>(output));
@@ -225,66 +107,15 @@ public class StreamSourceOperatorTest {
 		Assert.assertTrue(output.get(i).isWatermark());
 	}
 
-	@Test
-	public void testAutomaticWatermarkContext() throws Exception {
-
-		// regular stream source operator
-		final StoppableStreamSource<String, InfiniteSource<String>> operator =
-			new StoppableStreamSource<>(new InfiniteSource<String>());
-
-		long watermarkInterval = 10;
-		TestProcessingTimeService processingTimeService = new TestProcessingTimeService();
-		processingTimeService.setCurrentTime(0);
-
-		setupSourceOperator(operator, TimeCharacteristic.IngestionTime, watermarkInterval, 0, processingTimeService);
-
-		final List<StreamElement> output = new ArrayList<>();
-
-		StreamSourceContexts.getSourceContext(TimeCharacteristic.IngestionTime,
-			operator.getContainingTask().getProcessingTimeService(),
-			operator.getContainingTask().getCheckpointLock(),
-			operator.getContainingTask().getStreamStatusMaintainer(),
-			new CollectorOutput<String>(output),
-			operator.getExecutionConfig().getAutoWatermarkInterval(),
-			-1);
-
-		// periodically emit the watermarks
-		// even though we start from 1 the watermark are still
-		// going to be aligned with the watermark interval.
-
-		for (long i = 1; i < 100; i += watermarkInterval)  {
-			processingTimeService.setCurrentTime(i);
-		}
-
-		assertTrue(output.size() == 9);
-
-		long nextWatermark = 0;
-		for (StreamElement el : output) {
-			nextWatermark += watermarkInterval;
-			Watermark wm = (Watermark) el;
-			assertTrue(wm.getTimestamp() == nextWatermark);
-		}
-	}
-
 	// ------------------------------------------------------------------------
 
 	@SuppressWarnings("unchecked")
 	private static <T> void setupSourceOperator(StreamSource<T, ?> operator,
-			TimeCharacteristic timeChar,
-			long watermarkInterval,
-			long latencyMarkInterval) {
-		setupSourceOperator(operator, timeChar, watermarkInterval, latencyMarkInterval, new TestProcessingTimeService());
-	}
-
-	@SuppressWarnings("unchecked")
-	private static <T> void setupSourceOperator(StreamSource<T, ?> operator,
-												TimeCharacteristic timeChar,
-												long watermarkInterval,
-												long latencyMarkInterval,
-												final ProcessingTimeService timeProvider) {
+	                                            TimeCharacteristic timeChar,
+	                                            long latencyMarkInterval,
+	                                            final ProcessingTimeService timeProvider) {
 
 		ExecutionConfig executionConfig = new ExecutionConfig();
-		executionConfig.setAutoWatermarkInterval(watermarkInterval);
 		executionConfig.setLatencyTrackingInterval(latencyMarkInterval);
 
 		StreamConfig cfg = new StreamConfig(new Configuration());
@@ -321,40 +152,6 @@ public class StreamSourceOperatorTest {
 	}
 
 	// ------------------------------------------------------------------------
-
-	private static final class FiniteSource<T> implements SourceFunction<T>, StoppableFunction {
-
-		@Override
-		public void run(SourceContext<T> ctx) {}
-
-		@Override
-		public void cancel() {}
-
-		@Override
-		public void stop() {}
-	}
-
-	private static final class InfiniteSource<T> implements SourceFunction<T>, StoppableFunction {
-
-		private volatile boolean running = true;
-
-		@Override
-		public void run(SourceContext<T> ctx) throws Exception {
-			while (running) {
-				Thread.sleep(20);
-			}
-		}
-
-		@Override
-		public void cancel() {
-			running = false;
-		}
-
-		@Override
-		public void stop() {
-			running = false;
-		}
-	}
 
 	private static final class ProcessingTimeServiceSource implements SourceFunction<Long> {
